requirements.txt
segment-anything
nest-asyncio
numpy
torch
torchvision
opencv-python
Pillow
pycocotools
python-telegram-bot
gradio
huggingface_hub